RoleBinding for user for UI testing
